GenomePermutations

Documentation for GenomePermutations.

Index

GenomePermutations.AbstractGenomeDist

GenomePermutations.AdaptiveRegions

GenomePermutations.CircularRandomiser

GenomePermutations.PermTestResult

GenomePermutations.RegionRandomiser

GenomePermutations.SimplePTest

GenomePermutations.StartMixture

GenomePermutations.lengthsMixture

Base.rand

GenomePermutations._loopdist

GenomePermutations._randomgenome

GenomePermutations._randominterval

GenomePermutations.alloverlapping

GenomePermutations.anyin

GenomePermutations.anyoverlapping

GenomePermutations.countoverlapping

GenomePermutations.countoverlaps

GenomePermutations.dist

GenomePermutations.featuredist

GenomePermutations.generatedistribution

GenomePermutations.isin

GenomePermutations.iter_getcollection

GenomePermutations.overlappermtest

GenomePermutations.permtest

GenomePermutations.permtest

GenomePermutations.randominterval

GenomePermutations.randomise

GenomePermutations.randomisegenome

GenomePermutations.randomiseregions

GenomePermutations.simpleP

GenomePermutations.vec_getcollection

HypothesisTests.pvalue

HypothesisTests.pvalue

Introduction

Often we require to calculate how far or how many overlaps a series of genomic locations have with a set of predetermined features. We can naively compare our sample with other samples or a range of randomly selected locations along the genome. However, the distribution of overlaps or distances depends on the lengths of the intervals

Permutation tests, where we shuffle random regions to a create a random null distribution can provide a probabilistic framework. By generating a null distribution from intervals of the same length and distribution as what we observe, we can examine the location of our observation in the distribution to see how likely our value was due to chance alone.

GenomePermutations is an higher-performance - although there is still lots of space for optimization - opinionated Julia reimplementation of the permutation approach of RegioneR that aims to improve ease of use, composability and performance without sacrificing statistical rigor.

GenomePermutations package is designed with composability in mind: There are 2 main steps in running a permutation test

instantiate a GenomeDistribution type, subtype of AbstractGenomeDist, which contains both the regions, the new distribution and defines methods to draw random samples

define a distance function - sensible defaults are provided

run a permutation test.

Quick Start and Vignette

inspired by R's vignettes, this section will introduce a complete workflow, from loading bed files, to running the permutation test and visualizing the output

At the moment this is an unregistered package, so we need to provide a GitHub link to manually install it.

Note this will install from the main

using Pkg

Pkg.add(url="TBD")

if you want the nightly, most unstable and up to date you'll have to install from the dev branch

We then load our BED files with BED.jl

using GenomePermutations 
using BED
using plots

CNAs =  open(BED.Reader, "CNAs.bed") do reader
    IntervalCollection(reader, true)
end

features =  open(BED.Reader, "features.bed") do reader
    IntervalCollection(reader, true)
end

regions =  open(BED.Reader, "regions.bed") do reader
    IntervalCollection(reader, true)
end

Calculating distances:

distances = dist(CNAs, features)

Generating a genome distributions:

dist = GenomePermutations.StartMixture("hgTest", regions, false)

Running a permutation test:

test = permtest(CNAs,  features, dist, 1000)

# get the P value
pvalue(test.test)

histogram(test.ran, bins = :scott, color_palette=["grey75"]) # Ok, not the prettiest but it can fly for now

Checking if intervals are in a collection

These functions extend functions in GenomicFeatures to check if collections are contained in intervals or vice versa.

GenomePermutations.isin — Function

isin(a::GenomicFeatures.Interval{S}, GenomicFeatures.Interval{T})

Check if interval a is fully contained in interval b.

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 5, 10)
b = GenomicFeatures.Interval("chr1", 1, 15)

isin(a, b) 
	
# output

true

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 20, 30)
b = GenomicFeatures.Interval("chr1", 1, 15)

isin(a, b) 
	
# output

false

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 10, 20)
b = GenomicFeatures.Interval("chr1", 1, 15)

isin(a, b) 
	
# output

false

See Also GenomePermutations.isin, GenomePermutations.anyin.

isin(a::GenomicFeatures.Interval{T}, b::GenomicFeatures.IntervalCollection{S})

Linearly check if interval a is fully contained in any interval of collection b.

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 5, 10)
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

isin(a, b) 
		
# output

true

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 30, 50)
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35),
	GenomicFeatures.Interval("chr1", 35, 65)
	])

isin(a, b) 
		
# output

false

#Note

isin(a, b) strictly checks if a is contained in any single interval of b, rather than whether a is contained within any combination of intervals in collection b

See Also GenomePermutations.isin, GenomePermutations.anyin.

isin(a:::GenomicFeatures.IntervalCollection{T}, b:::GenomicFeatures.IntervalCollection{S})

Linearly check if all intervals in collection a are contained in collection b.

```jldoctest

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 6, 9),
	GenomicFeatures.Interval("chr1", 45, 50)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

[isin(a, b), isin(b, a)]

# output

2-element Vector{Bool}:
 true
 false
```

Note

At this point there are 2 layers of linear search so the time complexity is n^2.

See Also GenomePermutations.isin, GenomePermutations.anyin.

GenomePermutations.anyin — Function

anyin(a:::GenomicFeatures.IntervalCollection{T}, b:::GenomicFeatures.IntervalCollection{S})

Linearly check if any interval in collection a are contained in collection b.

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 25, 27),
	GenomicFeatures.Interval("chr1", 100, 150)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])
	
anyin(a, b)
	
# output

true

Note

At this point there are 2 layers of linear search so the time complexity is n^2

See Also GenomePermutations.isin, GenomePermutations.anyin.

Overlap calculations

GenomePermutations contains functions to calculate overlaps between intervals and/or collections

GenomePermutations.anyoverlapping — Function

anyoverlapping(a::GenomicFeatures.Interval{T},
b::GenomicFeatures.IntervalCollection{S})

Extend GenomicFeatures.isoverlapping to linearly check if interval a. overlaps collection. Return true or false

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 0, 10)
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 15),
	GenomicFeatures.Interval("chr1", 20, 25)
	])
anyoverlapping(a, b) 

# output

true

See Also GenomePermutations.alloverlapping, GenomePermutations.countoverlapping.

GenomePermutations.alloverlapping — Function

alloverlapping(a::GenomicFeatures.IntervalCollection{T}, b::GenomicFeatures.IntervalCollection{S})

Linearly check if all intervals in collection a overlap collection b. Return true or false

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 14, 17),
	GenomicFeatures.Interval("chr1", 20, 25)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 20, 25)
	])
alloverlapping(a, b) 
	
# output

false

See Also GenomePermutations.anyoverlapping, GenomePermutations.countoverlapping.

GenomePermutations.countoverlapping — Function

countoverlapping(a::GenomicFeatures.IntervalCollection{T} , b::GenomicFeatures.IntervalCollection{T})

Linearly count how many intervals in collection a overlap with any interval in collection b.

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 14, 17),
	GenomicFeatures.Interval("chr1", 20, 50)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

[countoverlapping(a, b), countoverlapping(b, a)]

# output

2-element Vector{Int64}:
 1
 2

#Note As seen above, countoverlapping(a,b) != countoverlapping(b,a) as one interval in a can overlap with >1 intervals in b.

See Also GenomePermutations.anyoverlapping, GenomePermutations.alloverlapping, GenomePermutations.countoverlaps.

GenomePermutations.countoverlaps — Function

countoverlaps(a::GenomicFeatures.IntervalCollection{T},b::GenomicFeatures.IntervalCollection{S})

linearly counts how many intervals in collection b overlaps with collection a

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 14, 17),
	GenomicFeatures.Interval("chr1", 20, 50)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

[countoverlaps(a, b), countoverlaps(b, a)]

# output

2-element Vector{Int64}:
 2
 2

#Note As seen above, countoverlapping(a,b) == countoverlapping(b,a) as we count the total number of overlaps for each interval

See Also GenomePermutations.anyoverlapping, GenomePermutations.alloverlapping, GenomePermutations.countoverlaps

Distance calculations

GenomePermutations contains functions to calculate distances between intervals and/or collections.

GenomePermutations.dist — Function

dist(a, b)

Return the unsigend distance between the two breakpoints of Intervals a and b or missing if the two intervals do not have the same seqname.

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 10, 20)
b = GenomicFeatures.Interval("chr1", 30, 65)
    
[dist(a, b), dist(b, a)]

# output

2-element Vector{Int64}:
 10
 10

See Also GenomePermutations.featuredist

GenomePermutations.featuredist — Function

featuredist(interval feature)

returns the feature distance between the breakpoints of an interval

Unilke normal distance, the feature distance takes into account overlaps between intervals and figures, returning 0 if an interval breakpoint

See Also GenomePermutations.dist

**NOTE:* Distance functions uses the maximum integer representable in memory to intialise the distance for performance reasons. However, in the unlikely event the real distance is equal to this value the function will return missing.

Genome Distributions

All Genome Distributions have shared methods.

All implement two methods, one to draw a random interval, and another to randomize a given interval.

The main difference is while randomise will use the length of the provided interval (and if by_chromosome = true, the chromosome) to, rand will always draw them at random from the lengths and chromosome distribution. Moreover, while rand will always return empty metadata with nothing type for metadata, randomise should conserve type and content of the metadata from the original - although this depends on the exact implementation of the specific genome distribution type

The following methods that extend rand, and randomise along a whole collection are shared by all Genome Distributions, unless otherwise specified.

Base.rand — Method

Base.rand(distribution::AbstractGenomeDist, n::Int)

Wraps the rand(<:AbstractGenomeDit) function to return a collection with nothing as metadata of n random intervals

Returns an interval collection with n intervals drawn from the distribution.

Note

if n = 1, a collection of size 1 is returned, if n = 0 a

GenomePermutations.randomise — Method

randomise(collection::IntervalCollection, distribution::AbstractGenomeDist)

Waps randomise(<:AbstractGenomeDist) to randomise a given IntervalCollection

Start Mixture

The core idea behind the Start Mixture genome distribution is that we can create a distribution of the all possible locations an interval can start and then check at sampling time if an interval of given length l can start at that location and still be contained in the allowed regions. If so, and optionally after checking it is not overlapping the nascient new interval list if the the interval is valid

GenomePermutations.StartMixture — Type

StartMixture(genome, regions, overlaps = false)

generates a StartsMixture

Under the hood it creates a ?

#arguments

#Avaibale fields - genome::String

Work in progress

Currently, there are some Genome Distributions - including some lifted straight out of RegioneR - that are sketched out but not yet implemented.

GenomePermutations.CircularRandomiser — Type

To Do The idea here is to rotate all segments by a x amount as if the chromosome was circular.

GenomePermutations.RegionRandomiser — Type

To Do

the idea here is to IDK what the idea is.

GenomePermutations.AdaptiveRegions — Type

To Do

the idea here is to regenerate the possible distribution at each iteartion

GenomePermutations.lengthsMixture — Type

To Do

the idea here is to create a series of possible start position for each length in a given collection

if any sounds like the kind, I'm always looking for contributors. See the section on extending Genome Permutations for more info

Permutation tests

Before we discuss the permutation test function, We need to introduce the structure that holds the results

GenomePermutations.PermTestResult — Type

PermTestResult{S <: Union{Real, Vector{<:Real}}, N <: Union{Real, Vector{<:Real}}, T <: Any}

A structure to store results of any permutation test implements pvalue.

HypothesisTests.pvalue — Method

HypothesisTests.pvalue(test::GenomePermutations.PermTestResult)

Extends HypothesisTests.pvalue to get the P value of a GenomePermutations.PermTestResult.

In the spirit of RegioneR, we can derive a quick P-value by simply

GenomePermutations.simpleP — Function

simpleP(obs, ran, iterations; alternative :Auto)

Calculates the p value for the alternative hypothesis that the observed value is either more or less than what we expect from the Vecor of random values.

If no altrantive is provided, will be automatically selected on the basis of the data.

See also: regioneR, SimplePTest Please see regioneR for full details

HypothesisTests.pvalue — Method

HypothesisTests.pvalue(test::GenomePermutations.SimplePTest)

Extends HypothesisTests.pvalue to get the P value of a GenomePermutations.SimplePTest.

GenomePermutations.SimplePTest — Type

SimplePTest(iterations::Int, p_val::, alternative::)

A structure to save RegioneR - style simple P tests. Implements pvalue

We can then run a permutation test with permtest

GenomePermutations.permtest — Method

permtest(randomised_collection, tested_collection, distribution, iterations, f; bed_names)

Runs a flexible permuation test, using a custsom evaluation function f.

Arguments

col1::GenomicFeatures.IntervalCollection{}: the collection to randomise.

col2::GenomicFeatures.IntervalCollection{}: the constant collection to test.

distribution::GenomicFeatures.AbstractGenomeDist: the genbome dist to use for the randomisation

iterations::Int: the number of iterations to run.

f::Function = GenomePermutations.dist: the evaluation function.

Defaults to GenomePermutations.dist.

bed_names::Tuple{Union{String, Nothing}}: the names of the bed files to write to.

Extending Genome Permutations

While sensible distance functions and genomeDistributions are provided, It should be possible to add new custom GenomDist types.

These custom types must be a subtype of the abstract AbstractGenomeDist type and must conform to the specification below

GenomePermutations.AbstractGenomeDist — Type

md AbstractGenomeDist is a base type for all AbstractGenomeDist objects.

All AbstractGenomeDist objects should have:

AT LEAST the following properties:

- genome::String the name of the genome  
- _regions::IntervalCollection the regions the genome distribution is  
- _distribution::Distribution the distribution to draw samples from  
- overlaps::Bool whether the sampled regions can overlap  
- _seqs the sequences of the regions used for the randomisation  
- _chr_distribution::Distribution the distribution to draw chromosomes from  
- _length_distribution::Distribution the distribution to draw lengths from  
- on_fail::Symbol. The action to take if the randomisation fails to draw a valid interval
shared methods recognise throw (throws an erorr), :continue (skips the interval) or :orig (returns the original interval)

note properties starting with _ are PRIVATE as they are derived from the regions used, please DO NOT ALLOW USERS TO SET THEM DIRECTLY. if you want to allow users to edit any of them you need to allow them to genereate all of them in one go with something like setfield!(self, :regions, value).

You must Implement the follwing methods:

- an appropriate constructor. should enforce on_fail to be one of the standard options, unless specialised methods to handle them are implemented  
- rand(): a method that returns a random interval from the distribution  
- randomise(): a method that returns a random interval from the distribution on the same chromosome
and with the same length of the provided interval

For reference, the following shared methods are available for all concrete subtypes of AbstractGenomeDistance:

Base.rand(::AbstractGenomeDist, ::Int)
randomise(::GenomicFeatures.IntervalCollection, ::AbstractGenomeDist)
Base.show(::AbstractGenomeDist)

It should also be possible to create a new distance function, as long as it returns 1 value (an Int, or missing) per interval we can rely on looping helpers to create appropriate versions that can handle collections, as needed

GenomePermutations._loopdist — Function

_loopdist(a, b, f)

_loopdist loops a distance function f that takes 2 intervals over a and b, handling corner cases.

Aeguments

a::GenomicFeatures.Interval or GenomicFeatures.IntervalCollection:

b::GenomicFeatures.IntervalCollection

f::Function

Note:

this is a private function that enables featuredist and dist to share the same loop and corner case handling core while being flexible enough to plug any function that takes two GenomicFeatures.Interval objects as arguments and returns an Int or missing

See Also GenomePermutations.dist, GenomePermutations.featuredist

_loopdist(a, b, f)

_loopdist loops a distance function f that takes 2 intervals over a and b, handling corner cases.

Aeguments

a::GenomicFeatures.Interval or GenomicFeatures.IntervalCollection:

b::GenomicFeatures.IntervalCollection

f::Function

Note:

this is a private function that enables featuredist and dist to share the same loop and corner case handling core while being flexible enough to plug any function that takes two GenomicFeatures.Interval objects as arguments and returns an Int or missing

See Also GenomePermutations.dist, GenomePermutations.featuredist

You can then alias your distance as needed

function mydist(a::Interval, b::IntervalCollection) 
    _loopdist(a, b, mydist)
end 

function  mydist(a::IntervalCollection, b::IntervalCollection)
    _loopdist(a, b, mydist)
end

in fact simply calling _loopdist from an untyped function should suffice.

function mydist(a, b) 
    _loopdist(a, b, mydist)
end

As long as you conform to both specifications, your new distance and genome distribution should work with the permutation test.

Legacy Functions

These functions were in use prior to the new Genome Distribution type systems are are deprecate due to their unstable and at times confusing output and clunky function calls. Will be officially deprecated and/or removed altogether by version 1.0.0

DEPRECATION WARNING!: PLEASE DO NOT USE THEM

GenomePermutations.generatedistribution — Function

generatedistribution(collection)

Generate a Distributions.MixtureModel from the collection

The model assumes each position of the genome has equal probability of being sampled. Practically, it is modelled as a mixture model of individual discrete uniform distributions, one for each interval in the collection. The component distributions are distribuited according to a categorical distribution, with probabilities proportional to the length of each interval.

Note

At the moment it only supports collections with single sequence. Generating a distribution from an overlapping collection is undocumented (I assume it would add up the probability but this is untested)

GenomePermutations.randominterval — Function

randominterval(interval, distribution, regions, collection; allow_overlap, max_tries, onfail)

Randomises the position of an interval according to a given distribution.

Arguments

interval::GenomicFeatures.IntervalCollection: The interval to randomise

distribution::Distributions.Sampleable: distribution of possible interval start location.

Should be a discrete distribution, although right now this is not enforced at the signature level

regions::: the regions the new interval should be fully contained in (checked byisin)

collecton::GenomicFeatures.IntervalCollection = GenomicFeatures.IntervalCollection{T}(): Collection to avoid overlapping

to if allow_overlap is set to false, defaults to an empty collection of metadata .

allow_overlap::Bool = true: whether to allow overlaps to collection.

max_tries::Int = 1000: the maximum number of attempts to draw an interval that fits in

the region from the distribution

onfail::Symbol = :throw: what to do if the interval fails to fit in the region.

Will return the orignal interval if :orig, returns nothing if :skip or throw an error if :throw.

Note

To prevent intervals in a loop from overlapping each other, use allow_overlap = false and add the result of each iteration to the collection paseed to the function.

See Also (_randominterval)[@ref]

GenomePermutations.randomisegenome — Function

randomisegenome(col, regions; allow_overlap, max_tries)

Randomises a collection, sequence by sequence along the specified regions, returning the new collection

rguments

col::GenomicFeatrues.IntervalCollection{T}: collection to randomise

regions::GenomicFeatures.IntervalCollection{S}: regions to randomise the collection in.

allow_overlap::Bool = false: whether to allow overlapping intervals in the result.

max_tries::Int = 1000: maximum number of attempts to draw an interval

that fits in the regions.

onfail::Symbol = :throw: what to do if an interval fails to fit in the regions.

Will return the orignal interval if :orig, skip the interval if :skip, otherwise throws an error.

Note: it will skip any sequences in the collection not found in the target regions. If no sequences are found returns an empty collection.

GenomePermutations.randomiseregions — Function

randomiseregions(collection, distribution, regions; allow_overlap, max_tries, onfail)

Generates randomised regions from an interval collection according to a start postion distribution. It guarantees that the new intervals are contained in the given regions.

Arguments

collection::GenomicFeatures.IntervalCollection{T}: collection to randomise,

must have only 1 sequence.

distribution::Distributions.Sampleable: distribution of possible interval

start locations, must match collection.

regions::GenomicFeatures.IntervalCollection{S}: regions to randomise the intervals in.

`allow_overlap::Bool = false.

max_tries::Int = 1000: maximum number of attempts to draw an interval

that fits in the region from the distribution.

onfail::Symbol = :throw: what to do if an interval fails to fit in the region.

Will return the orignal interval if :orig, skip the interval if :skip, otherwise throws an error.

Notes

this functions assumes that the interval collection contains intervals belonging to a single sequence that is the same sequence used for the distribution. The distribution should be derived from the regions. If the distribution and the regions don't match it will most likely fail safely New intervals overlapping with other intervals in the nascient collection are disallowed by default. To allow overlaps use `allow_overlap = true

GenomePermutations.overlappermtest — Function

overlappermtest(col1, col2, regions, iterations; bed_names, allow_overlap = false,
max_tries = 1000, onfail = :throws)

Legacy function to run an overlap permutation test across 2 regions. will be replaced by permtest using countoverlapping as f argument. Will need a g function to summarise overlaps across regions. (likely sum).

Arguments

col1::GenomicFeatures.IntervalCollection{T}: the collection to randomise.

col2::GenomicFeatures.IntervalCollection{S}: the constant collection to test.

regions::GenomicFeatures.IntervalCollection{S}: the regions col1 is randomised in.

iterations::Int: the number of iterations to run.

bed_names::Union{String, Nothing}: the names of the bed files to write to.

allow_overlap::Bool = false: whether to allow overlaps in the randomised regions.

max_tries::Int: the maximum number of tries to generate a randomised collection.

onfail::Symbol = :throw: what to do if a random interval fails to fit in the region.

Will return the orignal interval if :orig, skip the interval if :skip, otherwise throws an error.

GenomePermutations.permtest — Method

permtest(col1, col2, regions, iterations, f; allow_overlap = false, max_tries::Int = 1000, onfail = :throw)

Runs a flexible permuation test, using a custsom evaluation function f.

Arguments

col1::GenomicFeatures.IntervalCollection{}: the collection to randomise.

col2::GenomicFeatures.IntervalCollection{}: the constant collection to test.

regions::GenomicFeatures.IntervalCollection{}: the regions col1 is randomised in.

iterations::Int: the number of iterations torun.

f::Function = GenomePermutations.dist: the evaluation function.

Defaults to GenomePermutations.dist.

bed_names::TupleUnion{String, Nothing}: the names of the bed files to write to.

allow_overlap::Bool = false: whether to allow overlaps in the randomised regions.

max_tries::Int = 1000: the maximum number of tries to generate a randomised collection.

onfail::Symbol = :throw: what to do if a random interval fails to fit in the region.

Will return the orignal interval if :orig, skips it if :skip, otherwise throws an error.

These functions are not only deprecated but are also to be considered internal and private

GenomePermutations.iter_getcollection — Function

iter_getcollection(collection::GenomicFeatures.IntervalCollection{T}, sequence::String)

(inefficiently) retun a interval collection with only features in the specified sequence. There should be a better way to do this.

see also GenomePermutations.vec_getcollection

GenomePermutations.vec_getcollection — Function

vec_getcollection(collection::GenomicFeastures.IntervalCollection{T}, sequence::String)

(inefficiently) retun a interval collection with only features in the specified sequence. There should be a better way to do this.

See Also GenomePermutations.iter_getcollection

GenomePermutations._randomgenome — Function

_randomgenome(regions, n, C, L, names; allow_overlap = false)

Generates a new random genome with n segments from a named sequence distribution

Arguments

regions: the regions the genome to generate belongs to.

n::Int: the number of segments to generate.

C::Distributions.Categorical: a distibution with each cateogy representing

one of the named sequences. Lenght and order must match names.

`L::Distributions.DiscreteUniform": the segment lenght distribution.

names::Vector{Strimg}: the names of the sequences. must match the categories in C.

allow_overlap::Bool = false: whether to allow overlapping segments in the result.

This is currently private as I have not tested this and I don't like the categorical distribution squence names situation.

GenomePermutations._randominterval — Function

_randominterval(interval::GenomicFeatures.Interval{T}, 
	distribution::Distributions.Sampleable, regions::GenomicFeatures.IntervalCollection{S};
	collection::GenomicFeatures.IntervalCollection{T} = GenomicFeatures.IntervalCollection{T}(), 
	allow_overlap::Bool = true, max_tries::Int = 1000) where {T, S}}

This is the private version of randominterval, it skips sequence assertions in the name of performance Does it actually improve performance? Debeatable, in theory it prevents 100s of IFs. In practice, I haven't tested it. Indeed, branch prediction on asymmetrical branches should mean that the ifs are not really that damaging to performance.