GenomePermutations

Documentation for GenomePermutations.

GenomePermutations is an opinionated Julia reimplementation of R's RegioneR. Built on top of GenomicFeatures.jl it allows to genome permutation tests at scale

It additionally defines utility functions to calculate interval distances and overlaps, as well as

GenomePermutations.AbstractGenomeDist

GenomePermutations.AdaptiveRegions

GenomePermutations.CircularRandomiser

GenomePermutations.PermTestResult

GenomePermutations.RegionRandomiser

GenomePermutations.SimplePTest

GenomePermutations.StartMixture

GenomePermutations.lengthsMixture

GenomePermutations._randomgenome

GenomePermutations._randominterval

GenomePermutations.alloverlapping

GenomePermutations.anyin

GenomePermutations.anyoverlapping

GenomePermutations.countoverlapping

GenomePermutations.countoverlaps

GenomePermutations.dist

GenomePermutations.dist

GenomePermutations.dist

GenomePermutations.generatedistribution

GenomePermutations.isin

GenomePermutations.isin

GenomePermutations.isin

GenomePermutations.iter_getcollection

GenomePermutations.overlappermtest

GenomePermutations.permtest

GenomePermutations.randominterval

GenomePermutations.randomisegenome

GenomePermutations.randomiseregions

GenomePermutations.simpleP

GenomePermutations.vec_getcollection

HypothesisTests.pvalue

HypothesisTests.pvalue

GenomePermutations.AbstractGenomeDist — Type

AbstractGenomeDist is a base type for all AbstractGenomeDist objects.

All AbstractGenomeDist objects should have:

AT LEAST the following properties: - genome::String the name of the genome - regions::IntervalCollection the regions the genome distribution is - _distribution::Distribution the distribution to draw samples from - overlaps::Bool whether the sampled regions can overlap - _seqs the sequences of the regions used for the randomisation. - _chrdistribution::Distribution the distribution to draw chromosomes from - lengthdistribution::Distribution the distribution to draw lengths from - on_fail::Symbol. The action to take if the randomisation fails to draw a valid interval. shared methods recognise throw (throws an erorr), :continue (skips the interval) or :orig (returns the original interval)

note properties starting with _ are PRIVATE as they are derived from the regions used, pelase DO NOT SET THEM DIRECTLY. To edit any of them you need to alter the regions with setfield!(self, :regions, value).

Implement the follwing methods: - an appropriate constructor. should enforce on_fail to be one of the standard options, unless specialised methods to handle them are implemented. - setfields!(self, :regions, value) to change the regions - rand(): a method that returns a random interval from the distribution - randomise(): a method that returns a rnadom interval from the distribution on the same chromosome and with the same length of the provided interval.

AbstractGenomeDist has shared methods that can be used by all AbstractGenomeDist objects. - randomise(::AbstractGenomeDist, ): a method that wraps randomise() to randomise a given intervalcollection - rand(::AbstractGenomeDist, n::Int): a Wraps rand() to return a collection of n random intervals - pretty printing functions - getters and setteres please override amy abstract-level methods for custome genome rand objects as required

GenomePermutations.AdaptiveRegions — Type

To Do

the idea here is to regenerate the possible distribution at each iteartion

GenomePermutations.CircularRandomiser — Type

To Do The idea here is to rotate all segments by a x amount as if the chromosome was circular.

GenomePermutations.PermTestResult — Type

PermTestResult{S <: Union{Real, Vector{<:Real}}, N <: Union{Real, Vector{<:Real}}, T <: Any}

A structure to store results of any permutation test implements pvalue.

GenomePermutations.RegionRandomiser — Type

To Do

the idea here is to IDK what the idea is.

GenomePermutations.SimplePTest — Type

SimplePTest(iterations::Int, p_val::, alternative::)

A structure to save RegioneR - style simple P tests. Implements pvalue

GenomePermutations.StartMixture — Type

StartMixture(genome, regions, overlaps = false)

generates a genome rand object.

GenomePermutations.lengthsMixture — Type

To Do

the idea here is to create a series of possible start position for each length in a given collection

GenomePermutations._randomgenome — Method

_randomgenome(regions, n, C, L, names; allow_overlap = false)

Generates a new random genome with n segments from a named sequence distribution

Arguments

regions: the regions the genome to generate belongs to.

n::Int: the number of segments to generate.

C::Distributions.Categorical: a distibution with each cateogy representing

one of the named sequences. Lenght and order must match names.

`L::Distributions.DiscreteUniform": the segment lenght distribution.

names::Vector{Strimg}: the names of the sequences. must match the categories in C.

allow_overlap::Bool = false: whether to allow overlapping segments in the result.

This is currently private as I have not tested this and I don't like the categorical distribution squence names situation.

GenomePermutations._randominterval — Method

_randominterval(interval::GenomicFeatures.Interval{T}, 
	distribution::Distributions.Sampleable, regions::GenomicFeatures.IntervalCollection{S};
	collection::GenomicFeatures.IntervalCollection{T} = GenomicFeatures.IntervalCollection{T}(), 
	allow_overlap::Bool = true, max_tries::Int = 1000) where {T, S}}

This is the private version of randominterval, it skips sequence assertions in the name of performance Does it actually improve performance? Debeatable, in theory it prevents 100s of IFs. In practice, I haven't tested it. Indeed, branch prediction on asymmetrical branches should mean that the ifs are not really that damaging to performance.

GenomePermutations.alloverlapping — Method

alloverlapping(a::GenomicFeatures.IntervalCollection{T}, b::GenomicFeatures.IntervalCollection{S})

Linearly check if all intervals in collection a overlap collection b. Return true or false

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 14, 17),
	GenomicFeatures.Interval("chr1", 20, 25)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 20, 25)
	])
alloverlapping(a, b) 
	
# output

false

See Also GenomePermutations.anyoverlapping, GenomePermutations.countoverlapping.

GenomePermutations.anyin — Method

anyin(a:::GenomicFeatures.IntervalCollection{T}, b:::GenomicFeatures.IntervalCollection{S})

Linearly check if any interval in collection a are contained in collection b.

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 25, 27),
	GenomicFeatures.Interval("chr1", 100, 150)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])
	
anyin(a, b)
	
# output

true

Note

At this point there are 2 layers of linear search so the time complexity is n^2

See Also GenomePermutations.isin, GenomePermutations.anyin.

GenomePermutations.anyoverlapping — Method

anyoverlapping(a::GenomicFeatures.Interval{T},
b::GenomicFeatures.IntervalCollection{S})

Extend GenomicFeatures.isoverlapping to linearly check if interval a. overlaps collection. Return true or false

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 0, 10)
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 15),
	GenomicFeatures.Interval("chr1", 20, 25)
	])
anyoverlapping(a, b) 

# output

true

See Also GenomePermutations.alloverlapping, GenomePermutations.countoverlapping.

GenomePermutations.countoverlapping — Method

countoverlapping(a::GenomicFeatures.IntervalCollection{T} , b::GenomicFeatures.IntervalCollection{T})

Linearly count how many intervals in collection a overlap with any interval in collection b.

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 14, 17),
	GenomicFeatures.Interval("chr1", 20, 50)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

[countoverlapping(a, b), countoverlapping(b, a)]

# output

2-element Vector{Int64}:
 1
 2

#Note As seen above, countoverlapping(a,b) != countoverlapping(b,a) as one interval in a can overlap with >1 intervals in b.

See Also GenomePermutations.anyoverlapping, GenomePermutations.alloverlapping, GenomePermutations.countoverlaps.

GenomePermutations.countoverlaps — Method

countoverlaps(a::GenomicFeatures.IntervalCollection{T},b::GenomicFeatures.IntervalCollection{S})

linearly counts how many intervals in collection b overlaps with collection a

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 14, 17),
	GenomicFeatures.Interval("chr1", 20, 50)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

[countoverlapping(a, b), countoverlapping(b, a)]

# output

2-element Vector{Int64}:
 2
 2

#Note As seen above, countoverlapping(a,b) == countoverlapping(b,a) as we count the total number of overlaps for each interval

See Also GenomePermutations.anyoverlapping, GenomePermutations.alloverlapping, GenomePermutations.countoverlaps

GenomePermutations.dist — Method

dist(a,b, f = minimum)

Linearly calculate all the distances between an interval a and a collection b, returning a value summarised according to function f.

Arguments

a::GenomicFeatures.interval{T}: interval to calculate distances from.

b::GenomicFeatures.IntervalCollection{S}: collection to calculate all distances to.

f::Function: f Summarise all the pairwise distance for a and each interval in b

according to the supplied function f Currently tested with minimum, maximum, mean, and median.

using GenomicFeatures
using Statistics

a = GenomicFeatures.Interval("chr1", 25, 35)

b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 100)
	])

[dist(a,b, minimum), dist(a,b, maximum), dist(a,b, mean), dist(a,b, median)]
# output

4-element Vector{Float64}:
  0.0
 15.0
  6.666666666666667
  5.0

GenomePermutations.dist — Method

dist(a, b, f, g)

Lineraly caluclates the distance between collection a and b. Each distance between a and b is summarised according to f and then summarised according to g.

Agruments

a::GenomicFeatures.IntervalCollection{T}: collection a.

b::GenomicFeatures.IntervalCollection{S}: collectionb b.

f::Function = x -> x: f summarisees the result of the distances

between each interval of a and all intervals in b. tested with x -> x, minimum, 'maximum,mean,median`

g::Function = minimum: g summarises the distances returned by f for all

items in a. Currently tested with minimum. maximum, mean, and median.

Note

Both g and f can be replaced by user-defined functions. f must take a Vector{Union{Int, Float64}} and return Union{Int, Float64}. g must take a Vector{Union{Int, Float64}} but can return any type. Due to the fact that the result of f is stored in d before being summarised, it is not possible to change f's return type - even if we use a function in g that could acept a vector of that type. Indeed this is the reason for the need to Union{Int, Float64} as the return type of f. but types & their optimisation in Julia can be funny.

using GenomicFeatures
using Statistics

a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 15, 20),
	GenomicFeatures.Interval("chr1", 150, 200)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 100)
	])

[dist(a,b, x->x, minimum), dist(a,b, x->x, maximum), dist(a, x->x, mean), dist(a,b, x->x,median)]

# output

4-element Vector{Float64}:
  5.0
 50.0
 27.5
 27.5

GenomePermutations.dist — Method

dist(a, b)

Return the minimum unsigend distance between Intervals a and b. returns missing if the two intervals do not have the same seqname.

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 10, 20)
b = GenomicFeatures.Interval("chr1", 30, 65)
    
[dist(a, b), dist(b, a)]

# output

2-element Vector{Int64}:
 10
 10

GenomePermutations.generatedistribution — Method

generatedistribution(collection)

Generate a Distributions.MixtureModel from the collection

The model assumes each position of the genome has equal probability of being sampled. Practically, it is modelled as a mixture model of individual discrete uniform distributions, one for each interval in the collection. The component distributions are distribuited according to a categorical distribution, with probabilities proportional to the length of each interval.

Note

At the moment it only supports collections with single sequence. Generating a distribution from an overlapping collection is undocumented (I assume it would add up the probability but this is untested)

GenomePermutations.isin — Method

isin(a:::GenomicFeatures.IntervalCollection{T}, b:::GenomicFeatures.IntervalCollection{S})

Linearly check if all intervals in collection a are contained in collection b.

```jldoctest

using GenomicFeatures 
a = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 6, 9),
	GenomicFeatures.Interval("chr1", 45, 50)
	])
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

[isin(a, b), isin(b, a)]

# output

2-element Vector{Bool}:
 true
 false
```

Note

At this point there are 2 layers of linear search so the time complexity is n^2.

See Also GenomePermutations.isin, GenomePermutations.anyin.

GenomePermutations.isin — Method

isin(a::GenomicFeatures.Interval{T}, b::GenomicFeatures.IntervalCollection{S})

Linearly check if interval a is fully contained in any interval of collection b.

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 5, 10)
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35 ),
	GenomicFeatures.Interval("chr1", 40, 65)
	])

isin(a, b) 
		
# output

true

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 30, 50)
b = GenomicFeatures.IntervalCollection([
	GenomicFeatures.Interval("chr1", 5, 10),
	GenomicFeatures.Interval("chr1", 25, 35),
	GenomicFeatures.Interval("chr1", 35, 65)
	])

isin(a, b) 
		
# output

false

#Note

isin(a, b) strictly checks if a is contained in any single interval of b, rather than whether a is contained within any combination of intervals in collection b

See Also GenomePermutations.isin, GenomePermutations.anyin.

GenomePermutations.isin — Method

isin(a::GenomicFeatures.Interval{S}, GenomicFeatures.Interval{T})

Check if interval a is fully contained in interval b.

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 5, 10)
b = GenomicFeatures.Interval("chr1", 1, 15)

isin(a, b) 
	
# output

true

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 20, 30)
b = GenomicFeatures.Interval("chr1", 1, 15)

isin(a, b) 
	
# output

false

using GenomicFeatures 
a = GenomicFeatures.Interval("chr1", 10, 20)
b = GenomicFeatures.Interval("chr1", 1, 15)

isin(a, b) 
	
# output

false

See Also GenomePermutations.isin, GenomePermutations.anyin.

GenomePermutations.iter_getcollection — Method

iter_getcollection(collection::GenomicFeatures.IntervalCollection{T}, sequence::String)

(inefficiently) retun a interval collection with only features in the specified sequence. There should be a better way to do this.

see also GenomePermutations.vec_getcollection

GenomePermutations.overlappermtest — Method

overlappermtest(col1, col2, regions, iterations; bed_names, allow_overlap = false,
max_tries = 1000, onfail = :throws)

Legacy function to run an overlap permutation test across 2 regions. will be replaced by permtest using countoverlapping as f argument. Will need a g function to summarise overlaps across regions. (likely sum).

Arguments

col1::GenomicFeatures.IntervalCollection{T}: the collection to randomise.

col2::GenomicFeatures.IntervalCollection{S}: the constant collection to test.

regions::GenomicFeatures.IntervalCollection{S}: the regions col1 is randomised in.

iterations::Int: the number of iterations to run.

bed_names::Union{String, Nothing}: the names of the bed files to write to.

allow_overlap::Bool = false: whether to allow overlaps in the randomised regions.

max_tries::Int: the maximum number of tries to generate a randomised collection.

onfail::Symbol = :throw: what to do if a random interval fails to fit in the region.

Will return the orignal interval if :orig, skip the interval if :skip, otherwise throws an error.

GenomePermutations.permtest — Function

permtest(col1, col2, regions, iterations, f; allow_overlap = false, max_tries::Int = 1000, onfail = :throw)

Runs a flexible permuation test, using a custsom evaluation function f.

Arguments

col1::GenomicFeatures.IntervalCollection{}: the collection to randomise.

col2::GenomicFeatures.IntervalCollection{}: the constant collection to test.

regions::GenomicFeatures.IntervalCollection{}: the regions col1 is randomised in.

iterations::Int: the number of iterations to run.

f::Function = GenomePermutations.dist: the evaluation function.

Defaults to GenomePermutations.dist.

bed_names::Union{String, Nothing}: the names of the bed files to write to.

allow_overlap::Bool = false: whether to allow overlaps in the randomised regions.

max_tries::Int = 1000: the maximum number of tries to generate a randomised collection.

onfail::Symbol = :throw: what to do if a random interval fails to fit in the region.

Will return the orignal interval if :orig, skips it if :skip, otherwise throws an error.

GenomePermutations.randominterval — Method

randominterval(interval, distribution, regions, collection; allow_overlap, max_tries, onfail)

Randomises the position of an interval according to a given distribution.

Arguments

interval::GenomicFeatures.IntervalCollection: The interval to randomise

distribution::Distributions.Sampleable: distribution of possible interval start location.

Should be a discrete distribution, although right now this is not enforced at the signature level

regions::: the regions the new interval should be fully contained in (checked byisin)

collecton::GenomicFeatures.IntervalCollection = GenomicFeatures.IntervalCollection{T}(): Collection to avoid overlapping

to if allow_overlap is set to false, defaults to an empty collection of metadata .

allow_overlap::Bool = true: whether to allow overlaps to collection.

max_tries::Int = 1000: the maximum number of attempts to draw an interval that fits in

the region from the distribution

onfail::Symbol = :throw: what to do if the interval fails to fit in the region.

Will return the orignal interval if :orig, returns nothing if :skip or throw an error if :throw.

Note

To prevent intervals in a loop from overlapping each other, use allow_overlap = false and add the result of each iteration to the collection paseed to the function.

See Also (_randominterval)[@ref]

GenomePermutations.randomisegenome — Method

randomisegenome(col, regions; allow_overlap, max_tries)

Randomises a collection, sequence by sequence along the specified regions, returning the new collection

rguments

col::GenomicFeatrues.IntervalCollection{T}: collection to randomise

regions::GenomicFeatures.IntervalCollection{S}: regions to randomise the collection in.

allow_overlap::Bool = false: whether to allow overlapping intervals in the result.

max_tries::Int = 1000: maximum number of attempts to draw an interval

that fits in the regions.

onfail::Symbol = :throw: what to do if an interval fails to fit in the regions.

Will return the orignal interval if :orig, skip the interval if :skip, otherwise throws an error.

Note: it will skip any sequences in the collection not found in the target regions. If no sequences are found returns an empty collection.

GenomePermutations.randomiseregions — Method

randomiseregions(collection, distribution, regions; allow_overlap, max_tries, onfail)

Generates randomised regions from an interval collection according to a start postion distribution. It guarantees that the new intervals are contained in the given regions.

Arguments

collection::GenomicFeatures.IntervalCollection{T}: collection to randomise,

must have only 1 sequence.

distribution::Distributions.Sampleable: distribution of possible interval

start locations, must match collection.

regions::GenomicFeatures.IntervalCollection{S}: regions to randomise the intervals in.

`allow_overlap::Bool = false.

max_tries::Int = 1000: maximum number of attempts to draw an interval

that fits in the region from the distribution.

onfail::Symbol = :throw: what to do if an interval fails to fit in the region.

Will return the orignal interval if :orig, skip the interval if :skip, otherwise throws an error.

Notes

this functions assumes that the interval collection contains intervals belonging to a single sequence that is the same sequence used for the distribution. The distribution should be derived from the regions. If the distribution and the regions don't match it will most likely fail safely New intervals overlapping with other intervals in the nascient collection are disallowed by default. To allow overlaps use `allow_overlap = true

GenomePermutations.simpleP — Method

simpleP(obs, ran, iterations; alternative :Auto)

Calculates the p value for the alternative hypothesis that the observed value is either more or less than what we expect from the Vecor of random values.

If no altrantive is provided, will be automatically selected on the basis of the data.

See also: regioneR, SimplePTest Please see regioneR for full details

GenomePermutations.vec_getcollection — Method

vec_getcollection(collection::GenomicFeatures.IntervalCollection{T}, sequence::String)

(inefficiently) retun a interval collection with only features in the specified sequence. There should be a better way to do this.

See Also GenomePermutations.iter_getcollection

HypothesisTests.pvalue — Method

HypothesisTests.pvalue(test::GenomePermutations.PermTestResult)

Extends HypothesisTests.pvalue to get the P value of a GenomePermutations.PermTestResult.

HypothesisTests.pvalue — Method

HypothesisTests.pvalue(test::GenomePermutations.SimplePTest)

Extends HypothesisTests.pvalue to get the P value of a GenomePermutations.SimplePTest.